+++ ruby/spec/core_parity_spec.rb
@@ -43,7 +43,7 @@ RSpec.describe 'core parity (Ruby)' do
     expect { PdfOxide::PdfDocument.open('/no/such/file/does/not/exist.pdf') }.to raise_error(StandardError)
   end
 
-  it 'exposes version 0.3.71' do
-    expect(PdfOxide::VERSION).to eq('0.3.71')
+  it 'exposes version 0.3.72' do
+    expect(PdfOxide::VERSION).to eq('0.3.72')
   end
 end